Update stm32f4xx_hal_conf.h
The stm32f4xx_hal_conf.h HSE_VALUE definition doesn't work with micropython 1.11 while building the board file. It's missing (uint32_t) against a regex in pllvalues.py

Error: 
python3 boards/pllvalues.py -c  file:boards/BLACK_F407VE/stm32f4xx_hal_conf.h > build-BLACK_F407VE/genhdr/pllfreqtable.h
Traceback (most recent call last):
  File "boards/pllvalues.py", line 205, in <module>
    main()
  File "boards/pllvalues.py", line 178, in main
    raise ValueError("%s does not contain a definition of HSE_VALUE" % argv[0])
ValueError: file:boards/BLACK_F407VE/stm32f4xx_hal_conf.h does not contain a definition of HSE_VALUE
Makefile:565: recipe for target 'build-BLACK_F407VE/genhdr/pllfreqtable.h' failed
make: *** [build-BLACK_F407VE/genhdr/pllfreqtable.h] Error 1
make: *** Deleting file 'build-BLACK_F407VE/genhdr/pllfreqtable.h'

+++ stm32f4xx_hal_conf.h
@@ -4,7 +4,7 @@
 #include "boards/stm32f4xx_hal_conf_base.h"
 
 // Oscillator values in Hz
-#define HSE_VALUE (8000000)
+#define HSE_VALUE ((uint32_t)8000000)
 #define LSE_VALUE (32768)
 #define EXTERNAL_CLOCK_VALUE (12288000)
